# Backend/MainService/openapi.yaml
openapi: 3.1.0
info:
  title: User Service API
  version: 1.0.0
  description: OpenAPI спецификация для основного сервиса
paths:
...
  /api/v1/main_service/projects/{projectId}/comments:
    post:
      tags:
        - ProjectComment
      summary: Создание комментария к проекту
      description: Добавляет комментарий текущего пользователя к указанному проекту.
      operationId: createProjectComment
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '200':
          description: Комментарий успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
    get:
      tags:
        - ProjectComment
      summary: Получение комментариев проекта
      description: Возвращает список всех комментариев к проекту.
      operationId: getProjectComments
      parameters:
        - name: projectId
          in: path
          required: true
          description: Идентификатор проекта
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
  /api/v1/main_service/comments/{id}:
    delete:
      tags:
        - ProjectComment
      summary: Удаление комментария
      description: Удаляет комментарий с указанным идентификатором. Доступно только автору комментария.
      operationId: deleteComment
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор комментария
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Комментарий успешно удален
          content: { }
...
  /api/v1/main_service/news/{newsId}/comments:
    post:
      tags:
        - NewsComment
      summary: Создание комментария к новости
      description: Добавляет комментарий текущего пользователя к указанной новости.
      operationId: createNewsComment
      parameters:
        - name: newsId
          in: path
          required: true
          description: Идентификатор новости
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '200':
          description: Комментарий успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
    get:
      tags:
        - NewsComment
      summary: Получение комментариев новости
      description: Возвращает список всех комментариев к новости.
      operationId: getNewsComments
      parameters:
        - name: newsId
          in: path
          required: true
          description: Идентификатор новости
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
  /api/v1/main_service/news-comments/{id}:
    delete:
      tags:
        - NewsComment
      summary: Удаление комментария к новости
      description: Удаляет комментарий с указанным идентификатором. Доступно только автору комментария.
      operationId: deleteNewsComment
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор комментария
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Комментарий успешно удален
          content: { }
  /api/v1/main_service/news/{id}/likes:
    post:
      tags:
        - NewsLike
      summary: Лайк/дизлайк новости
      description: Переключает лайк текущего пользователя для указанной новости.
      operationId: toggleNewsLike
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор новости
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Состояние лайка после операции
          content:
            application/json:
              schema:
                type: boolean
                example: true
  /api/v1/main_service/news/{id}/likes/count:
    get:
      tags:
        - NewsLike
      summary: Получение количества лайков новости
      description: Возвращает текущее количество лайков новости.
      operationId: getNewsLikesCount
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор новости
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Количество лайков новости
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /api/v1/main_service/articles/{articleId}/comments:
    post:
      tags:
        - ArticleComment
      summary: Создание комментария к статье
      description: Добавляет комментарий текущего пользователя к указанной статье.
      operationId: createArticleComment
      parameters:
        - name: articleId
          in: path
          required: true
          description: Идентификатор статьи
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentRequest'
      responses:
        '200':
          description: Комментарий успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
    get:
      tags:
        - ArticleComment
      summary: Получение комментариев статьи
      description: Возвращает список всех комментариев к статье.
      operationId: getArticleComments
      parameters:
        - name: articleId
          in: path
          required: true
          description: Идентификатор статьи
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
  /api/v1/main_service/article-comments/{id}:
    delete:
      tags:
        - ArticleComment
      summary: Удаление комментария к статье
      description: Удаляет комментарий с указанным идентификатором. Доступно только автору комментария.
      operationId: deleteArticleComment
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор комментария
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Комментарий успешно удален
          content: { }
  /api/v1/main_service/articles/{id}/likes:
    post:
      tags:
        - ArticleLike
      summary: Лайк/дизлайк статьи
      description: Переключает лайк текущего пользователя для указанной статьи.
      operationId: toggleArticleLike
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор статьи
          schema:
            type: integer
            format: int64
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Состояние лайка после операции
          content:
            application/json:
              schema:
                type: boolean
                example: true
  /api/v1/main_service/articles/{id}/likes/count:
    get:
      tags:
        - ArticleLike
      summary: Получение количества лайков статьи
      description: Возвращает текущее количество лайков статьи.
      operationId: getArticleLikesCount
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор статьи
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Количество лайков статьи
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /api/v1/main_service/news-categories:
    get:
      tags:
        - NewsCategory
      summary: Получение всех категорий новостей
      description: Возвращает список всех категорий новостей
      operationId: getNewsCategories
      responses:
        '200':
          description: Список категорий новостей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsCategorySummaryResponse'
    post:
      tags:
        - NewsCategory
      summary: Создание категории новости
      description: Создание новой категории новости
      operationId: createNewsCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsCategoryRequest'
      responses:
        '201':
          description: Категория новости успешно создана
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/NewsCategoryFullResponse'
  /api/v1/main_service/news-categories/{id}:
    get:
      tags:
        - NewsCategory
      summary: Получение категории новости по ID
      description: Получение полной информации о категории новости по ее ID
      operationId: getNewsCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Категория новости найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsCategoryFullResponse'
        '404':
          description: Категория новости не найдена
          content:
            text/plain:
              schema:
                type: string
    patch:
      tags:
        - NewsCategory
      summary: Обновление категории новости
      description: Обновление данных категории новости по ее ID
      operationId: updateNewsCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsCategoryRequest'
      responses:
        '200':
          description: Категория новости успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsCategoryFullResponse'
        '404':
          description: Категория новости не найдена
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
        - NewsCategory
      summary: Удаление категории новости
      description: Удаление категории новости по ее ID
      operationId: deleteNewsCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Категория новости успешно удалена
        '404':
          description: Категория новости не найдена
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    NewsCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Technology
    NewsCategorySummaryResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Technology
    NewsCategoryFullResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Technology
...
    CommentRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Текст комментария
          minLength: 1
          example: "test"
...
    UserFullResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 42
        username:
          type: string
          description: Имя пользователя
          example: "Andrey"
...
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор комментария
          example: 1001
        author:
          $ref: '#/components/schemas/UserFullResponse'
        text:
          type: string
          description: Текст комментария
          example: "User12345"
        createdAt:
          type: string
          description: Дата и время создания комментария
          example: "2026-01-24T14:32:10"
...
  parameters:
    XUserNameHeader:
      name: X-User-Name
      in: header
      description: "Имя пользователя, проставляется Gateway"
      required: true
      schema:
        type: string